[feature] 레거시 MVC와 애노테이션 기반 MVC 통합 - HomeController 변경
