+++ src/main/java/next/controller/HomeController.java
@@ -1,15 +1,19 @@
 package next.controller;
 
-import core.db.DataBase;
-import core.mvc.asis.Controller;
+import core.annotation.web.Controller;
+import core.annotation.web.RequestMapping;
+import core.annotation.web.RequestMethod;
+import core.mvc.JspView;
+import core.mvc.ModelAndView;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class HomeController implements Controller {
-    @Override
-    public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
-        req.setAttribute("users", DataBase.findAll());
-        return "home.jsp";
+@Controller
+public class HomeController {
+
+    @RequestMapping(value = "/", method = RequestMethod.GET)
+    public ModelAndView home(HttpServletRequest req, HttpServletResponse resp) throws Exception {
+        return new ModelAndView(new JspView("home.jsp"));
     }
 }
