+++ src/main/java/next/controller/asis/ProfileController.java
@@ -1,4 +1,4 @@
-package next.controller;
+package next.controller.asis;
 
 import core.db.DataBase;
 import core.mvc.asis.Controller;
@@ -7,16 +7,15 @@ import next.model.User;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class UpdateFormUserController implements Controller {
-
+public class ProfileController implements Controller {
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         String userId = req.getParameter("userId");
         User user = DataBase.findUserById(userId);
-        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
-            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
+        if (user == null) {
+            throw new NullPointerException("사용자를 찾을 수 없습니다.");
         }
         req.setAttribute("user", user);
-        return "/user/updateForm.jsp";
+        return "/user/profile.jsp";
     }
 }
