+++ src/main/java/next/controller/asis/UpdateFormUserController.java
@@ -1,21 +1,23 @@
-package next.controller;
+package next.controller.asis;
 
 import core.db.DataBase;
 import core.mvc.asis.Controller;
+import next.controller.UserSessionUtils;
 import next.model.User;
 
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
-public class ProfileController implements Controller {
+public class UpdateFormUserController implements Controller {
+
     @Override
     public String execute(HttpServletRequest req, HttpServletResponse resp) throws Exception {
         String userId = req.getParameter("userId");
         User user = DataBase.findUserById(userId);
-        if (user == null) {
-            throw new NullPointerException("사용자를 찾을 수 없습니다.");
+        if (!UserSessionUtils.isSameUser(req.getSession(), user)) {
+            throw new IllegalStateException("다른 사용자의 정보를 수정할 수 없습니다.");
         }
         req.setAttribute("user", user);
-        return "/user/profile.jsp";
+        return "/user/updateForm.jsp";
     }
 }
